backend/requirements.txt
Django==5.2.6
psycopg[binary]==3.2.10
djangorestframework==3.15.2
django-environ==0.12.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1